requirements.txt
gradio==5.35.0
pillow
huggingface_hub==0.33.2
python-dotenv
google-generativeai==0.8.4
google-auth-oauthlib==1.2.1
google-auth==2.38.0
google-api-python-client==2.166.0
google-api-core==2.24.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-genai==1.8.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
openai
